Reset option window currently selected when refreshing list
If an OptionSelectWindow is refreshed, there is the possibility of
there being less items in the list than there were previously. Since
the "selected" option didn't change, that also opened the possibility
of the selected option being outside the range of visible
options. Now, when refreshing the list of options, reset the selected
option to the top of the list.